+++ CHANGELOG.md
@@ -4,7 +4,14 @@
 
 ### Features
 ### Improvements
+
+* Improve loading times for the rule tree by optimizing the rule segmentation and sorting.
+* Add timeout of 10 seconds to the loading time for each rule and make it configurable in the tree 
+configuration file via `rule_load_timeout`. Warn if a rule was not loaded due to this.
+
 ### Bugfix
+
+* Fix the rule tree parsing some rules incorrectly, potentially resulting in more matches.
 
 ## v7.0.0
 ### Breaking